requirements.txt
paho-mqtt==1.5.1
google-cloud-pubsub==3.8.0
